requirements.txt
# ===================================
# Requirements for BA-Agent Backend
# ===================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# LangChain & AI
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langgraph>=0.0.50
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# Vector Store
chromadb>=0.4.0

# Database
pymongo>=4.6.0

# Environment & Utils
python-dotenv>=1.0.0
pandas>=2.0.0

# Google AI
google-generativeai>=0.3.0
google-genai>=0.8.0
google-api-core>=2.15.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# JSON Processing
json_repair>=0.0.12

# Typing Extensions
typing_extensions>=4.5.0
